compose compact with pruning command to one
